+++ src/alpha_beta/debug.rs
@@ -2,7 +2,39 @@ use crate::Game;
 
 use std::fmt::{self, Debug};
 
-use super::{BestAction, MiniMax, State, Terminated};
+use super::{Action, Branch, Ctxt, MiniMax, State};
+
+impl<T: Game> Debug for Action<T>
+where
+    T::Action: Debug,
+    T::Fitness: Debug,
+{
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        f.debug_struct("Action")
+            .field("fitness", &self.fitness)
+            .field("path", &self.fitness)
+            .finish()
+    }
+}
+
+impl<'a, T: Game> Debug for Ctxt<'a, T>
+where
+    T: Debug,
+    T::Action: Debug,
+    T::Fitness: Debug,
+    T::Player: Debug,
+{
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        f.debug_struct("Ctxt")
+            .field("state", &self.state)
+            .field("player", &self.player)
+            .field("best", &self.best)
+            .field("unfinished", &self.unfinished)
+            .field("terminated", &self.terminated)
+            .field("partially_terminated", &self.partially_terminated)
+            .finish()
+    }
+}
 
 impl<T: Game> Debug for MiniMax<T>
 where
@@ -24,7 +56,7 @@ where
     T::Fitness: Debug,
 {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        f.debug_struct("BestAction")
+        f.debug_struct("State")
             .field("alpha", &self.alpha)
             .field("beta", &self.beta)
             .field("best_fitness", &self.best_fitness)
@@ -35,29 +67,16 @@ where
     }
 }
 
-impl<T: Game> Debug for BestAction<T>
+impl<T: Game> Debug for Branch<T>
 where
     T::Action: Debug,
     T::Fitness: Debug,
 {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        f.debug_struct("BestAction")
-            .field("action", &self.action)
-            .field("fitness", &self.fitness)
-            .field("path", &self.path)
-            .finish()
-    }
-}
-
-impl<T: Game> Debug for Terminated<T>
-where
-    T::Action: Debug,
-    T::Fitness: Debug,
-{
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        f.debug_struct("Terminated")
-            .field("best_action", &self.best_action)
-            .field("partial", &self.partial)
-            .finish()
+        match self {
+            Branch::Worse(fitness) => write!(f, "Worse({:?})", fitness),
+            Branch::Better(fitness) => write!(f, "Better({:?})", fitness),
+            Branch::Equal(fitness) => write!(f, "Equal({:?})", fitness),
+        }
     }
 }