+++ src/alpha_beta/mod.rs
@@ -3,7 +3,6 @@
 use crate::{Game, IntoRunCondition, RunCondition};
 
 use std::cmp;
-use std::fmt::{self, Debug};
 use std::mem;
 
 mod debug;
@@ -75,71 +74,54 @@ impl<T: Game> Bot<T> {
             return None;
         }
 
-        let mut actions = actions.into_iter().collect::<Vec<_>>();
-        if actions.len() < 2 {
-            return actions.pop();
-        }
-        actions.sort_by_cached_key(|a| state.look_ahead(&a, self.player));
+        let mut ctxt = Ctxt::new(
+            state,
+            self.player,
+            actions
+                .into_iter()
+                .map(|action| Action {
+                    fitness: state.look_ahead(&action, self.player),
+                    path: vec![action],
+                })
+                .collect(),
+        );
 
-        let mut terminated = Terminated::default();
-        let mut best_action: Option<BestAction<T>> = None;
+        // Guard against the trivial case.
+        if ctxt.unfinished.len() < 2 {
+            return ctxt.unfinished.pop().map(|mut act| act.path.pop().unwrap());
+        }
+
         for depth in 0.. {
             if !condition.depth(depth) {
-                return current_best(terminated, best_action).or_else(|| actions.pop());
+                return ctxt.cancel().path.pop();
             }
 
-            let prev_actions = mem::replace(&mut actions, Vec::new());
+            let mut unfinished = mem::take(&mut ctxt.unfinished);
+            // Try unfinished actions with a high expected fitness first,
+            // as they are expected to give us a better alpha value.
+            unfinished.sort_by_key(|act| act.fitness);
 
-            if let Some(BestAction {
-                path,
-                action,
-                fitness,
-            }) = best_action.take()
-            {
-                let alpha = terminated.best_fitness();
-
-                match self.rate_action(state, action, path, alpha, depth, &mut condition) {
-                    RateAction::Cancelled(action) => {
-                        return current_best(
-                            terminated,
-                            Some(BestAction {
-                                path: Vec::new(),
-                                action,
-                                fitness,
-                            }),
-                        )
-                    }
-                    RateAction::NewBest(new) => best_action = Some(new),
-                    RateAction::Worse(action) => actions.push(action),
-                    RateAction::UpperLimit(action) => return Some(action),
-                    RateAction::Terminated(action, fitness) => {
-                        terminated.add_complete(action, fitness)
-                    }
-                    RateAction::PartiallyTerminated(action, fitness) => {
-                        terminated.add_partial(action, fitness)
-                    }
+            if let Some(best) = ctxt.best.take() {
+                // If computation is cancelled here, we don't know anything new,
+                // so we can just return the previous best action.
+                if let Some(mut ret) = ctxt.try_action(best, depth, &mut condition, |_, act| act) {
+                    return ret.path.pop();
                 }
             }
 
-            for action in prev_actions.into_iter().rev() {
-                let alpha = alpha(&terminated, &best_action);
-                match self.rate_action(state, action, Vec::new(), alpha, depth, &mut condition) {
-                    RateAction::Cancelled(action) => {
-                        return current_best(terminated, best_action).or(Some(action))
+            for action in unfinished.into_iter().rev() {
+                // In case computation is cancelled here, we may not yet have computed the best action of
+                // the previous depth, to guard against this, we add the cancelled action back to `state.unfinished`
+                // in case it is still empty.
+                let on_cancel = |ctxt: &mut Ctxt<T>, act| {
+                    if ctxt.unfinished.is_empty() {
+                        ctxt.unfinished.push(act);
                     }
-                    RateAction::NewBest(new) => {
-                        if let Some(prev_best) = best_action.replace(new) {
-                            actions.push(prev_best.action)
-                        }
-                    }
-                    RateAction::Worse(action) => actions.push(action),
-                    RateAction::UpperLimit(action) => return Some(action),
-                    RateAction::Terminated(action, fitness) => {
-                        terminated.add_complete(action, fitness)
-                    }
-                    RateAction::PartiallyTerminated(action, fitness) => {
-                        terminated.add_partial(action, fitness)
-                    }
+                    ctxt.cancel()
+                };
+
+                if let Some(mut ret) = ctxt.try_action(action, depth, &mut condition, on_cancel) {
+                    return ret.path.pop();
                 }
             }
 
@@ -149,79 +131,266 @@ impl<T: Game> Bot<T> {
             // As the current best fitness does not come from a terminated path,
             // we still have to keep the other partially terminated actions around,
             // in case the best fitness of a later depth is lower.
-            for action in
-                terminated.relevant_partials(best_action.as_ref().map(|best| best.fitness))
-            {
-                let alpha = alpha(&terminated, &best_action);
-                match self.rate_action(state, action, Vec::new(), alpha, depth, &mut condition) {
-                    RateAction::Cancelled(_action) => return current_best(terminated, best_action),
-                    RateAction::NewBest(new) => {
-                        if let Some(prev_best) = best_action.replace(new) {
-                            actions.push(prev_best.action)
-                        }
-                    }
-                    RateAction::Worse(action) => actions.push(action),
-                    RateAction::UpperLimit(action) => return Some(action),
-                    RateAction::Terminated(action, fitness) => {
-                        terminated.add_complete(action, fitness)
-                    }
-                    RateAction::PartiallyTerminated(action, fitness) => {
-                        terminated.add_partial(action, fitness)
-                    }
+            for action in ctxt.relevant_partials() {
+                // In case computation is cancelled here, we already tested at least some actions which were better than
+                // the cancelled partial action at the previous depth, so we can use `ctxt.cancel()` without any special
+                // considerations.
+                if let Some(mut ret) =
+                    ctxt.try_action(action, depth, &mut condition, |ctxt, _| ctxt.cancel())
+                {
+                    return ret.path.pop();
                 }
             }
 
             // All partially terminated actions are worse than all completely terminated actions.
-            if actions.is_empty() && best_action.is_none() {
-                assert!(terminated.partial.is_empty());
-                return Some(terminated.best_action.map(|(a, _f)| a).unwrap());
+            if ctxt.unfinished.is_empty() && ctxt.best.is_none() {
+                assert!(ctxt.partially_terminated.is_empty());
+                return ctxt.terminated.unwrap().path.pop();
             }
         }
 
         unreachable!();
     }
+}
 
-    fn rate_action<U: RunCondition>(
-        &self,
-        state: &T,
-        action: T::Action,
-        path: Vec<T::Action>,
-        alpha: Option<T::Fitness>,
+#[derive(Clone, Copy, Debug)]
+struct CancelledError;
+
+enum MiniMax<T: Game> {
+    /// No new elements were found in this branch
+    Terminated(Vec<T::Action>, Branch<T>),
+    /// New elements were found
+    Open(Vec<T::Action>, Branch<T>),
+    /// There are no possible actions for this state
+    DeadEnd,
+}
+
+impl<T: Game> MiniMax<T> {
+    /// Appends an action to self.
+    pub fn with(self, action: T::Action, fitness: T::Fitness) -> MiniMax<T> {
+        match self {
+            MiniMax::DeadEnd => MiniMax::Terminated(vec![action], Branch::Equal(fitness)),
+            MiniMax::Open(mut actions, branch) => {
+                actions.push(action);
+                MiniMax::Open(actions, branch)
+            }
+            MiniMax::Terminated(mut actions, branch) => {
+                actions.push(action);
+                MiniMax::Terminated(actions, branch)
+            }
+        }
+    }
+}
+
+enum Branch<T: Game> {
+    /// `actual_fitness <= fitness`
+    Worse(T::Fitness),
+    /// `actual_fitness >= fitness`
+    Better(T::Fitness),
+    /// `actual_fitness == fitness`
+    Equal(T::Fitness),
+}
+
+impl<T: Game> Branch<T> {
+    fn fitness(&self) -> T::Fitness {
+        match self {
+            Branch::Worse(fitness) | Branch::Better(fitness) | Branch::Equal(fitness) => *fitness,
+        }
+    }
+}
+
+/// A top level action.
+struct Action<T: Game> {
+    /// The current fitness of a given action.
+    ///
+    /// This can mean one of the following things:
+    ///
+    /// - For the best unfinished action, this is exact, but only for the current depth.
+    /// - For a terminated action, this is exact.
+    /// - For a partially terminated action, this is the upper limit.
+    fitness: T::Fitness,
+    /// The expected path taken during optimal play, when only  inspectnig up to the current depth.
+    ///
+    /// This is ordered with `path.remove(0)` being the first action.
+    path: Vec<T::Action>,
+}
+
+/// The currently available data at the highest level, during minimax `State` is used instead.
+pub struct Ctxt<'a, T: Game> {
+    /// The initial gamestate.
+    state: &'a T,
+    /// The maximizing player.
+    player: T::Player,
+    /// The best unfinished action. This is not set if there is an already better terminated action.
+    best: Option<Action<T>>,
+    /// Actions which are both not yet finished and worse than `best_unfinished`.
+    unfinished: Vec<Action<T>>,
+    /// The best already completely terminated action. We keep the path for diagnostic
+    /// purposes only, as there is no reason to retry this.
+    terminated: Option<Action<T>>,
+    /// Partially terminated actions. These are paths which had a cutoff at the highest level.
+    ///
+    /// As these actions cannot have a fitness higher than this cutoff, we discard all partially terminated
+    /// actions which must be worse than `best_terminated`.
+    partially_terminated: Vec<Action<T>>,
+}
+
+impl<'a, T: Game> Ctxt<'a, T> {
+    fn new(state: &T, player: T::Player, unfinished: Vec<Action<T>>) -> Ctxt<T> {
+        Ctxt {
+            state,
+            player,
+            best: None,
+            unfinished,
+            terminated: None,
+            partially_terminated: Vec::new(),
+        }
+    }
+
+    /// Returns all partially terminated actions may be better than `self.best_unfinished`,
+    /// and should therefore be retried at the current depth.
+    fn relevant_partials(&mut self) -> impl IntoIterator<Item = Action<T>> {
+        self.partially_terminated.sort_by_key(|act| act.fitness);
+
+        if let Some(ref best) = self.best {
+            // We only care about partially terminated paths which may be better than the current best.
+            let pos = self
+                .partially_terminated
+                .iter()
+                .position(|act| act.fitness > best.fitness)
+                .unwrap_or(self.partially_terminated.len());
+            self.partially_terminated.split_off(pos)
+        } else {
+            mem::take(&mut self.partially_terminated)
+        }
+    }
+
+    /// Updates `self.terminated` in case the new action has a higher fitness.
+    ///
+    /// This also removes all partially terminated actions with a worse maximum fitness,
+    /// as they are now irrelevant.
+    fn add_terminated(&mut self, act: Action<T>) {
+        if self
+            .terminated
+            .as_ref()
+            .map_or(true, |best| best.fitness < act.fitness)
+        {
+            self.partially_terminated
+                .retain(|partial| partial.fitness > act.fitness);
+            // We only keep `best` if it is better than `terminated`.
+            if let Some(best) = self.best.take() {
+                if best.fitness > act.fitness {
+                    // Still relevant, put it back in.
+                    self.best = Some(best);
+                } else {
+                    // Not relevant, add it to the other unfinished actions.
+                    self.unfinished.push(best);
+                }
+            }
+
+            self.terminated = Some(act);
+        }
+    }
+
+    /// Adds a new partially finished action in case its maximum fitness is
+    /// greater than the fitness of the best completely terminated action.
+    fn add_partially_terminated(&mut self, act: Action<T>) {
+        if self
+            .terminated
+            .as_ref()
+            .map_or(true, |best| best.fitness < act.fitness)
+        {
+            self.partially_terminated.push(act);
+        }
+    }
+
+    fn add_best(&mut self, act: Action<T>) {
+        if self
+            .best
+            .as_ref()
+            .or(self.terminated.as_ref())
+            .map_or(true, |best| best.fitness < act.fitness)
+        {
+            // Move the previous best action back into `unfinished`.
+            self.unfinished.extend(self.best.replace(act));
+        } else {
+            self.unfinished.push(act);
+        }
+    }
+
+    /// Stop computing and return the currently best action.
+    fn cancel(&mut self) -> Action<T> {
+        self.best
+            .take()
+            .or(self.terminated.take())
+            .unwrap_or_else(|| {
+                mem::take(&mut self.unfinished)
+                    .into_iter()
+                    .max_by_key(|act| act.fitness)
+                    .unwrap()
+            })
+    }
+
+    /// Tests the given action at the current depth, returns `Some`
+    /// once we are finished.
+    fn try_action<U: RunCondition>(
+        &mut self,
+        action: Action<T>,
         depth: u32,
         condition: &mut U,
-    ) -> RateAction<T> {
-        let mut updated_state = state.clone();
-        let fitness = updated_state.execute(&action, self.player);
-        match self.minimax_with_path(path, updated_state, depth, alpha, None, condition) {
-            Err(CancelledError) => RateAction::Cancelled(action),
+        on_cancel: impl FnOnce(&mut Self, Action<T>) -> Action<T>,
+    ) -> Option<Action<T>> {
+        let mut updated_state = self.state.clone();
+        let mut path = action.path.clone();
+
+        let fitness = updated_state.execute(&path.pop().unwrap(), self.player);
+        match self.minimax_with_path(
+            path.into_iter().rev(),
+            updated_state,
+            depth,
+            self.best
+                .as_ref()
+                .or(self.terminated.as_ref())
+                .map(|act| act.fitness),
+            None,
+            condition,
+        ) {
+            Err(CancelledError) => Some(on_cancel(self, action)),
             Ok(MiniMax::DeadEnd) => {
-                if state.is_upper_bound(fitness, self.player) {
-                    RateAction::UpperLimit(action)
+                if self.state.is_upper_bound(fitness, self.player) {
+                    Some(action)
                 } else {
-                    RateAction::Terminated(action, fitness)
+                    self.add_terminated(action);
+                    None
                 }
             }
-            Ok(MiniMax::Terminated(_path, Branch::Equal(fitness))) => {
-                if state.is_upper_bound(fitness, self.player) {
-                    RateAction::UpperLimit(action)
+            Ok(MiniMax::Terminated(mut path, Branch::Equal(fitness))) => {
+                path.push(action.path.last().unwrap().clone());
+                let action = Action { fitness, path };
+                if self.state.is_upper_bound(fitness, self.player) {
+                    Some(action)
                 } else {
-                    RateAction::Terminated(action, fitness)
+                    self.add_terminated(action);
+                    None
                 }
             }
-            Ok(MiniMax::Terminated(_path, Branch::Worse(maximum))) => {
-                RateAction::PartiallyTerminated(action, maximum)
+            Ok(MiniMax::Terminated(mut path, Branch::Worse(fitness))) => {
+                path.push(action.path.last().unwrap().clone());
+                let action = Action { fitness, path };
+                self.add_partially_terminated(action);
+                None
             }
-            Ok(MiniMax::Open(_path, Branch::Worse(_))) => RateAction::Worse(action),
-            Ok(MiniMax::Open(path, Branch::Equal(fitness))) => {
-                if Some(fitness) > alpha {
-                    RateAction::NewBest(BestAction {
-                        action,
-                        fitness,
-                        path,
-                    })
-                } else {
-                    RateAction::Worse(action)
-                }
+            Ok(MiniMax::Open(mut path, Branch::Worse(_))) => {
+                path.push(action.path.last().unwrap().clone());
+                let action = Action { fitness, path };
+                self.unfinished.push(action);
+                None
+            }
+            Ok(MiniMax::Open(mut path, Branch::Equal(fitness))) => {
+                path.push(action.path.last().unwrap().clone());
+                let action = Action { fitness, path };
+                self.add_best(action);
+                None
             }
             Ok(MiniMax::Terminated(_, Branch::Better(_)))
             | Ok(MiniMax::Open(_, Branch::Better(_))) => {
@@ -261,7 +430,7 @@ impl<T: Game> Bot<T> {
     /// we very quickly get a good alpha/lower limit.
     fn minimax_with_path<U: RunCondition>(
         &self,
-        mut path: Vec<T::Action>,
+        mut path: impl Iterator<Item = T::Action>,
         game_state: T,
         depth: u32,
         alpha: Option<T::Fitness>,
@@ -272,7 +441,7 @@ impl<T: Game> Bot<T> {
             return Err(CancelledError);
         }
 
-        let action = if let Some(action) = path.pop() {
+        let action = if let Some(action) = path.next() {
             action
         } else {
             return self.minimax(game_state, depth, alpha, beta, condition);
@@ -364,66 +533,6 @@ impl<T: Game> Bot<T> {
         }
 
         Ok(state.consume())
-    }
-}
-
-#[derive(Clone, Copy, Debug)]
-struct CancelledError;
-
-enum MiniMax<T: Game> {
-    /// No new elements were found in this branch
-    Terminated(Vec<T::Action>, Branch<T>),
-    /// New elements were found
-    Open(Vec<T::Action>, Branch<T>),
-    /// There are no possible actions for this state
-    DeadEnd,
-}
-
-impl<T: Game> MiniMax<T> {
-    /// Appends an action to self.
-    pub fn with(self, action: T::Action, fitness: T::Fitness) -> MiniMax<T> {
-        match self {
-            MiniMax::DeadEnd => MiniMax::Terminated(vec![action], Branch::Equal(fitness)),
-            MiniMax::Open(mut actions, branch) => {
-                actions.push(action);
-                MiniMax::Open(actions, branch)
-            }
-            MiniMax::Terminated(mut actions, branch) => {
-                actions.push(action);
-                MiniMax::Terminated(actions, branch)
-            }
-        }
-    }
-}
-
-enum Branch<T: Game> {
-    /// `actual_fitness <= fitness`
-    Worse(T::Fitness),
-    /// `actual_fitness >= fitness`
-    Better(T::Fitness),
-    /// `actual_fitness == fitness`
-    Equal(T::Fitness),
-}
-
-impl<T: Game> Branch<T> {
-    fn fitness(&self) -> T::Fitness {
-        match self {
-            Branch::Worse(fitness) | Branch::Better(fitness) | Branch::Equal(fitness) => *fitness,
-        }
-    }
-}
-
-impl<T: Game> Debug for Branch<T>
-where
-    T::Action: Debug,
-    T::Fitness: Debug,
-{
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            Branch::Worse(fitness) => write!(f, "Worse({:?})", fitness),
-            Branch::Better(fitness) => write!(f, "Better({:?})", fitness),
-            Branch::Equal(fitness) => write!(f, "Equal({:?})", fitness),
-        }
     }
 }
 
@@ -538,21 +647,19 @@ impl<T: Game> State<T> {
                     self.update_best_action(path, Branch::Equal(fitness));
                 }
             }
+        } else if terminated && self.state.is_lower_bound(fitness, self.player) {
+            self.alpha = Some(fitness);
+            self.beta = Some(fitness);
+            self.best_fitness = Some(Branch::Equal(fitness));
+            self.terminated = true;
         } else {
-            if terminated && self.state.is_lower_bound(fitness, self.player) {
-                self.alpha = Some(fitness);
-                self.beta = Some(fitness);
-                self.best_fitness = Some(Branch::Equal(fitness));
-                self.terminated = true;
-            } else {
-                self.beta = Some(self.beta.map_or(fitness, |value| cmp::min(value, fitness)));
-                if self
-                    .best_fitness
-                    .as_ref()
-                    .map_or(true, |old| old.fitness() >= fitness)
-                {
-                    self.update_best_action(path, Branch::Equal(fitness));
-                }
+            self.beta = Some(self.beta.map_or(fitness, |value| cmp::min(value, fitness)));
+            if self
+                .best_fitness
+                .as_ref()
+                .map_or(true, |old| old.fitness() >= fitness)
+            {
+                self.update_best_action(path, Branch::Equal(fitness));
             }
         }
     }
@@ -600,127 +707,6 @@ impl<T: Game> State<T> {
             MiniMax::Terminated(self.path, self.best_fitness.unwrap())
         } else {
             MiniMax::Open(self.path, self.best_fitness.unwrap())
-        }
-    }
-}
-
-struct BestAction<T: Game> {
-    action: T::Action,
-    fitness: T::Fitness,
-    path: Vec<T::Action>,
-}
-
-/// Contains data about already terminated paths.
-struct Terminated<T: Game> {
-    /// The fitness of the best completely finished action.
-    best_action: Option<(T::Action, T::Fitness)>,
-    /// Actions which terminated due to a cutoff, meaning that `fitness >= actual fitness`.
-    partial: Vec<(T::Action, T::Fitness)>,
-}
-
-impl<T: Game> Default for Terminated<T> {
-    #[inline]
-    fn default() -> Self {
-        Terminated {
-            best_action: None,
-            partial: Vec::new(),
-        }
-    }
-}
-
-impl<T: Game> Terminated<T> {
-    /// Returns all partially terminated actions which might be better than `best_fitness`.
-    #[inline]
-    fn relevant_partials(&mut self, best_fitness: Option<T::Fitness>) -> Vec<T::Action> {
-        let mut relevant = Vec::new();
-        for (action, fitness) in mem::replace(&mut self.partial, Vec::new()) {
-            if Some(fitness) > best_fitness {
-                relevant.push(action);
-            } else {
-                self.partial.push((action, fitness));
-            }
-        }
-        relevant
-    }
-
-    /// Updates `self.best_action` in case the new action has a higher fitness.
-    ///
-    /// This also removes all partially terminated actions with a worse maximum fitness,
-    /// as they are now irrelevant.
-    fn add_complete(&mut self, action: T::Action, fitness: T::Fitness) {
-        if self.best_fitness().map_or(true, |best| best < fitness) {
-            self.best_action = Some((action, fitness));
-            self.partial.retain(|(_a, f)| *f > fitness);
-        }
-    }
-
-    /// Adds a new partially finished action in case its maximum fitness is
-    /// greater than the fitness of the best completely terminated action.
-    fn add_partial(&mut self, action: T::Action, fitness: T::Fitness) {
-        if self.best_fitness().map_or(true, |best| best < fitness) {
-            self.partial.push((action, fitness));
-        }
-    }
-
-    /// returns the fitness of the best completely terminated action
-    fn best_fitness(&self) -> Option<T::Fitness> {
-        self.best_action.as_ref().map(|(_a, fitness)| *fitness)
-    }
-}
-
-fn current_best<T: Game>(
-    terminated: Terminated<T>,
-    best_action: Option<BestAction<T>>,
-) -> Option<T::Action> {
-    match (terminated.best_action, best_action) {
-        (Some(term), Some(best)) => Some(if best.fitness > term.1 {
-            best.action
-        } else {
-            term.0
-        }),
-        (Some(term), None) => Some(term.0),
-        (None, Some(best)) => Some(best.action),
-        (None, None) => None,
-    }
-}
-
-#[inline]
-fn alpha<T: Game>(
-    terminated: &Terminated<T>,
-    best_action: &Option<BestAction<T>>,
-) -> Option<T::Fitness> {
-    cmp::max(
-        best_action.as_ref().map(|best| best.fitness),
-        terminated.best_fitness(),
-    )
-}
-
-enum RateAction<T: Game> {
-    Cancelled(T::Action),
-    NewBest(BestAction<T>),
-    Worse(T::Action),
-    UpperLimit(T::Action),
-    Terminated(T::Action, T::Fitness),
-    PartiallyTerminated(T::Action, T::Fitness),
-}
-
-impl<T: Game> Debug for RateAction<T>
-where
-    T::Action: Debug,
-    T::Fitness: Debug,
-{
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            RateAction::Cancelled(action) => write!(f, "Cancelled({:?})", action),
-            RateAction::NewBest(best_action) => write!(f, "NewBest({:?})", best_action),
-            RateAction::Worse(action) => write!(f, "Worse({:?})", action),
-            RateAction::UpperLimit(action) => write!(f, "UpperLimit({:?})", action),
-            RateAction::Terminated(action, fitness) => {
-                write!(f, "Terminated({:?}, {:?})", action, fitness)
-            }
-            RateAction::PartiallyTerminated(action, fitness) => {
-                write!(f, "PartiallyTerminated({:?}, {:?})", action, fitness)
-            }
         }
     }
 }